:bug: Fixes empty cells being ignored when fetching data
The airtable.js library ignores empty cells of each record and doesn't provide their keys either. That resulted in the arrays being created from each record to be different in size, depending on the number of empty cells. To mitigate that, we now set the value of empty cells to be an empty string. That also fixes the issue of values being shifted to the wrong index.