:bug: Fixes bug with module's types and order of properties
The type bug prevented typescript projects using the integration from compiling. Besides that, Airtable API doesn't provide the table data in correct order, which this commit fixes by adding a config option called order, that allows the user to specify the order themselves. The framework handles that issue for the ResponseTable in the background