:sparkles: overhauls amazon pay integration. details in next paragraph
We added an AmazonPay module that offers payload builders for the setup and charge directive. Because of that we renamed the AmazonPay service file to AmazonPayAPI. We also added a feature to map the alexa api to the amazon pay api. Because of that, the host property in the options of amazon pay api requests is not necessary anymore.