:bug: the autopilot output functions always add to beginning of actions arr
Currently we have implemented the say, play, listen, and show actions. These should always be at the beginning of the actions array. The rest of the actions which are can be accessed using setAction(), will always be at the end. Otherwise we run into the risk of having an invalid sequence. This might have to be changed if we decide to add the rest of the actions to the framework