:boom: :recycle: saves session ID in user.session object. Changes session object structure
some platforms do not provide an property that specifies if its a new session or not in the request. We save the session ID in the database to compare it with the request's one to determine if its a new session. To make the implementation cleaner, we added it to the user object's session object. For that we had to change its structure since it just stored the data of session.data at that time