:white_check_mark: Autopilot: switch to FileDb2 and remove randomUserId()
using FileDb with a static user ID for each of the request resulted in a corrupt db.json file. Made the tests fail sometimes. To fix that we switched to FileDb2, i.e. each user has their own file, and add a random user ID to each request by default. Because of that we also deleted the randomUserId stuff from the test files. It's done by the RequestBuilder now