:sparkles: :bug: Implemented check for duplicated messages
Google Business Messages sends multiple messages with the same `messageId` if the response is not returned fast enough. For example, this can happen in a Lambda when it warms up or if time-intensive api-calls are made.
In order to avoid handling a message twice, the `messageId` will be stored in the session-data of the user and checked for after setting the session-data. An immediate save is required because a Lambda may start multiple instances which then need to be aware of the request from the other instance.