:label: Improve signature of `t` and `$t`
If the i18n resource is linked in the typing-system, it is possible to automatically retrieve the correct type when using `t` and `$t`. If that fails for whatever reason, the return type is the same as the base return type of `i18next.t`. Forcing the result type is still possible via the first generic parameter.