alexa-verifier-middleware v 2.x is not supported
Version 2.x is now a pure es module, and requires node 12.17 or higher. If you want to run this via an older version of node, use alexa-verifier-middleware@1.x
If we change to version 2.x we also need to use an import statement instead of require.
It would also make sense to add it to the dependencies to have control over min/max version numbers.
And i prefer to finally throw an error here. Would have saved me 1hrs of debugging.